+++ plugins/action/icap_spectrum_interference_device_reports_info.py
@@ -2,8 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Copyright (c) 2021, Cisco Systems
-# GNU General Public License v3.0+ (see LICENSE or
-# https://www.gnu.org/licenses/gpl-3.0.txt)
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
 
@@ -32,8 +31,8 @@ argument_spec.update(
         startTime=dict(type="float"),
         endTime=dict(type="float"),
         apMac=dict(type="str"),
-        limit=dict(type="float"),
-        offset=dict(type="float"),
+        limit=dict(type="int"),
+        offset=dict(type="int"),
         timeSortOrder=dict(type="str"),
         headers=dict(type="dict"),
     )
